+++ changelog.md
@@ -1,4 +1,4 @@
-# [UNRELEASED] - 2025-09-25
+# [UNRELEASED] - 2025-09-29
 
 ## Added 
 - Added register account
@@ -7,8 +7,11 @@
 - Added list parametric
 - Added set defaults
 - Added service list type parametrics
+- Added service homologate products
+- Added service list homologated products
 
 ## Changed
 - Changed use token implementation in services
 - Changed parametric service instance in emizorapi implementation
 - Changed type parametrics, validation
+- Change validation parametric type